requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
langchain-nvidia-ai-endpoints>=0.3.0
langchain-community>=0.3.0
nemoguardrails>=0.10.0
faiss-cpu>=1.8.0
streamlit>=1.38.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
requests>=2.32.0
